+++ examples/imgviz/src/args.rs
@@ -30,7 +30,139 @@
 */
 
 use std::path::PathBuf;
+
+use bpaf::{construct, long, short, OptionParser, Parser};
 use tiny_skia::Color;
+
+#[allow(dead_code)]
+#[derive(Debug, Clone)]
+pub(crate) struct Out {
+    pub(crate) debug: bool,
+    pub(crate) in_filename: PathBuf,
+    pub(crate) out_filename: PathBuf,
+    pub(crate) resolution: u32,
+    pub(crate) side: Option<u8>,
+    pub(crate) sides: Option<u8>,
+    pub(crate) hole_ratio: f32,
+    pub(crate) angle: f32,
+    pub(crate) data: bool,
+    pub(crate) weak: bool,
+    pub(crate) errors: bool,
+    pub(crate) metadata: bool,
+    pub(crate) index_hole: bool,
+    pub(crate) decode: bool,
+    pub(crate) cc: bool,
+    pub(crate) supersample: u32,
+    pub(crate) img_bg_color: Option<Color>,
+    pub(crate) track_bg_color: Option<Color>,
+    pub(crate) title: Option<String>,
+}
+
+/// Set up bpaf argument parsing.
+pub(crate) fn opts() -> OptionParser<Out> {
+    let debug = short('d').long("debug").help("Print debug messages").switch();
+
+    let in_filename = short('i')
+        .long("in_filename")
+        .help("Filename of disk image to read")
+        .argument::<PathBuf>("IN_FILE");
+
+    let out_filename = short('o')
+        .long("out_filename")
+        .help("Filename of image to write")
+        .argument::<PathBuf>("OUT_FILE");
+
+    let resolution = short('r')
+        .long("resolution")
+        .help("Size of resulting image, in pixels")
+        .argument::<u32>("SIZE");
+
+    let side = short('s')
+        .long("side")
+        .help("Side to render. Omit to render both sides")
+        .argument::<u8>("SIDE")
+        .guard(|side| *side < 2, "Side must be 0 or 1")
+        .optional();
+
+    let sides = long("sides")
+        .help("Override number of sides to render. Only useful for rendering single-sided disks as double-wide images.")
+        .argument::<u8>("SIDES")
+        .guard(|sides| *sides > 0 && *sides < 3, "Sides must be 1 or 2")
+        .optional();
+
+    let hole_ratio = short('h')
+        .long("hole_ratio")
+        .help("Ratio of inner radius to outer radius")
+        .argument::<f32>("RATIO")
+        .fallback(0.33);
+
+    let angle = short('a')
+        .long("angle")
+        .help("Angle of rotation")
+        .argument::<f32>("ANGLE")
+        .fallback(0.0);
+
+    let data = long("data").help("Render data").switch();
+
+    let weak = short('w').long("weak").help("Render weak bits").switch();
+
+    let errors = short('e').long("errors").help("Render bitstream errors").switch();
+
+    let metadata = long("metadata").help("Render metadata").switch();
+
+    let decode = long("decode").help("Decode data").switch();
+
+    let index_hole = long("index_hole").help("Render index hole").switch();
+
+    let supersample = long("ss")
+        .help("Supersample (2,4,8)")
+        .argument::<u32>("FACTOR")
+        .fallback(1);
+
+    let cc = long("cc").help("Wrap data counter-clockwise").switch();
+
+    let img_bg_color = long("img_bg_color")
+        .help("Specify the image background color as #RRGGBBAA, #RRGGBB, or R,G,B,A")
+        .argument::<String>("IMAGE_BACKGROUND_COLOR")
+        .parse(|input: String| parse_color(&input))
+        .optional();
+
+    let track_bg_color = long("track_bg_color")
+        .help("Specify the track background color as #RRGGBBAA, #RRGGBB, or R,G,B,A")
+        .argument::<String>("TRACK_BACKGROUND_COLOR")
+        .parse(|input: String| parse_color(&input))
+        .optional();
+
+    // Title argument with substitution
+    let title = long("title")
+        .help("Specify the title string, or ${IN_FILE} to use the input filename.")
+        .argument::<String>("TITLE")
+        .optional();
+
+    construct!(Out {
+        debug,
+        in_filename,
+        out_filename,
+        resolution,
+        side,
+        sides,
+        hole_ratio,
+        angle,
+        data,
+        weak,
+        errors,
+        metadata,
+        index_hole,
+        decode,
+        cc,
+        supersample,
+        img_bg_color,
+        track_bg_color,
+        title,
+    })
+    .to_options()
+    .descr("imgviz: generate a graphical visualization of a disk image")
+}
 
 /// Perform `${IN_FILE}` substitution for the `title`, using only the filename portion of `in_filename`.
 pub(crate) fn substitute_title(title: Option<String>, in_filename: &PathBuf) -> Option<String> {